+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+Fixes:
+
+- Fixes #11 - make `FailureResult#response` public
+
 ## [0.5.3][] (2021-11-03)
 
 Fixes: